SCRUM-207 refactor: move payment return-flow logic into feature layer
