+++ src/features/profile/settings/index.ts
@@ -1,3 +1,2 @@
-export { AccountManagement } from './ui/AccountManagement'
 export { Devices } from './ui/Devices'
 export { GeneralInfo } from './ui/General/GeneralInfo'
