+++ src/features/subscriptions/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './usePaymentsTable'
+export * from './usePaymentReturnFlow'
