+++ src/features/subscriptions/ui/AccountManagement/index.ts
@@ -1,0 +1,1 @@
+export * from './AccountManagement'
