+++ src/features/subscriptions/ui/index.ts
@@ -1,2 +1,3 @@
-export * from './SubscriptionPricing'
 export * from './Payments'
+export * from './AccountManagement'
+export * from './SubscriptionPricing'
